HTML Symbols & Display:inline
I replaced the greater and less than symbols three times with three different kinds of html symbols (this is becuase it clashed with the html tag notation that uses them). I also retained 2 paragraphs with display:inline style to show how it works (I previously used it to prevent the clash with '<' and '>' but realized it was too messy...insteady I use symbols that each start with '&' and end with ';').