<Svg> and <Rect> tags added
I added the <svg> and <rect> tags to make a red square (I added the autofocus attribute to the <svg> tags so the page loads on the element which helps with debugging in this long pieces of code). I also added the height, width, and fill attributes to the rectangle class.