<svg> with <feOffset> and <feBlend> added
I added an <svg> with <feOffset> and <feBlend> elements that helped create the effect of 2 blocks on top of each other (these are between the elements <filter> and <defs>). <feOffset> element's "dx" and "dy" attributes indicate how far to move the square below it relative from the top and left sides of the square on top of it.