Top-left link to page's top with "position: fixed" added
I added a blue <div> element in the top-left corner with "position: fixed" so I can see it at any point when I scroll down. I also made sure it was a link to allow me to go back to the top of the page (this is allowed because I have a set of <a> tags around the words in the <div> element).  This acts like a home button.