Home icon added to "homeButton" class
I added a home icon using the class "fa fa-home" to the div with the "homeButton" class that provides a link to the top regardless of how far down you scroll. I added a link element with the href attribute pointing to the Font Awesome icon library website to allow me to use the home icon.