Immediate child css relationship using > symbol
I added a css element that uses the relationship ">" to indicate that only paragraphs that are an immediate child of details are to have a lightgray css property as noted by (details > p). I added a corresponding paragraph within a div to show how only that paragraph isn't an immediate child since it's between the div tags.