General Sibling Selector added
I added the general sibling selector so that any paragraphs located below the time selector that are children (aka not wrapped around other elements) will be highlighted with the background color "peachpuff". I added a third paragraph that the rule is applied to despite it being located towards the bottom of the html document past the other elements.