"Transform: translate(50px, 50px)" added
I added the "transform: translate(50px, 50px)" property so it will move the div 50px to the right and 50px down respectively. I added this with three other similar properties but with the "-ms-", "-webkit-", and "-moz" prefixes so it will work completely with the older versions of the internet explorer/edge, chrome/safari/opera, and firefox browsers respectively.