Added three divs to position peach pictures
I added three divs where the outer div contains all the peach pictures and have "position: relative" with a set width and height. This helps contain the other 2 divs that each cover about half the pictures which have "position: absolute" so i could put half of the pictures in a second column (using "left: 300px" to move the div to the left of the one with normal positioning) to save space.