Moved homebutton link to sidebar
I moved the div containing a link to the top of the screen inside the div associated with the sidebar. I also changed the homebutton's position to "relative" so it's position is relative to the sidebar and gave it a width of "100%" so it will fill the width of the sidebar. This is to ensure the homebutton doesn't cover up the other elements when I click on links from the sidebar to go to different sections.